refactor community board components
